dialogue: move playing logic to control
