+++ projects/plugins/jetpack/extensions/blocks/dialogue/edit.js
@@ -19,7 +19,7 @@ import {
 	ToolbarButton,
 } from '@wordpress/components';
 import { useContext, useState, useEffect, useLayoutEffect, useRef } from '@wordpress/element';
-import { useSelect, useDispatch, dispatch } from '@wordpress/data';
+import { useSelect, dispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -36,7 +36,6 @@ import {
 	list as defaultParticipants,
 } from '../conversation/participants.json';
 import { formatUppercase } from '../../shared/icons';
-import { STORE_ID as MEDIA_SOURCE_STORE_ID, STORE_ID } from '../../store/media-source/constants';
 import MediaPlayerControl from '../../shared/components/media-player-control';
 
 function getParticipantBySlug( participants, slug ) {
@@ -78,32 +77,10 @@ export default function DialogueEdit( {
 	const richTextRef = useRef();
 	const baseClassName = 'wp-block-jetpack-dialogue';
 
-	const {
-		prevBlock,
-		playerState,
-		playerCurrentTime,
-		defaultMediaSource,
-	} = useSelect( select => {
-		// Pick the previous block atteobutes from the state.
+	const prevBlock = useSelect( select => {
 		const prevPartClientId = select( 'core/block-editor' ).getPreviousBlockClientId( clientId );
-
-		// Media source:
-		const {
-			getMediaSourceCurrentTime,
-			getMediaPlayerState,
-			getDefaultMediaSource,
-		} = select( MEDIA_SOURCE_STORE_ID );
-
-		return {
-			prevBlock: select( 'core/block-editor' ).getBlock( prevPartClientId ),
-			playerState: getMediaPlayerState(),
-			playerCurrentTime: getMediaSourceCurrentTime(),
-			defaultMediaSource: getDefaultMediaSource(),
-		}
+		return select( 'core/block-editor' ).getBlock( prevPartClientId );
 	}, [] );
-
-	const { toggleMediaSource } = useDispatch( STORE_ID );
-	const togglePlayer = () => toggleMediaSource( defaultMediaSource.id );
 
 	// Block context integration.
 	const participantsFromContext = context[ 'jetpack/conversation-participants' ];
@@ -232,15 +209,10 @@ export default function DialogueEdit( {
 	return (
 		<div className={ className }>
 			<BlockControls>
-				{ defaultMediaSource && (
-					<MediaPlayerControl
-						state={ playerState }
-						time={ playerCurrentTime }
-						skipForwardTime={ false }
-						jumpBackTime={ false }
-						onToggle={ togglePlayer }
-					/>
-				) }
+				<MediaPlayerControl
+					jumpBackTime={ false }
+					skipForwardTime={ false }
+				/>
 
 				{ currentParticipant && isFocusedOnParticipantLabel && (
 					<ToolbarGroup>
