+++ projects/plugins/jetpack/extensions/shared/components/media-player-control/Readme.md
@@ -1,33 +1,9 @@
 Media Player control
 ====================
 
-Implement a media player control.
-
-```es6
-function MyMediaComponent() {
-	const [ playerState, setPlayerState ] = useState( STATE_PAUSED );
-
-	<MediaPlayerControl
-		state={ playerState }
-		onToggle={ setPlayerState }
-	/>
-}
-```
+Implement a media player control connected to the [Media Source Store](./extensions/stores/media-source/).
 
 ## Props
-
-### time [s]
-Player current time, in seconds.
-
-* Type: `Number`
-* Required: No
-
-### state
-Player current state. It could take `is-playing` or `is-paused`.
-We encourage using constants defined in the [media source store](./extensions/stores/media-source/constants.js), instead of custom strings.
-
-* Type: `String`
-* Required: No
 
 ### skipForwardTime
 Time value to skip forward when clicking on the forwarding button. Or, `False` to do not render the button. `five` seconds as default.
@@ -57,25 +33,6 @@ Icon to show in the `pause` button. `controls-pause` as default.
 * Type: `?(string|WPElement)`
 * Required: No
 * Default: `controls-pause`
-
-### isDisabled
-Disables/Enables all player buttons.
-
-* Type: `boolean`
-* Required: No
-* Default: `False`
-
-### onTimeChange
-A callback invoked when the time value changes. It passes time (number) as a callback parameter.
-
-* Type: `Function`
-* Required: No
-
-### onToggle
-A callback invoked when the player state changes. It passes state (string) as a callback parameter.
-
-* Type: `Function`
-* Required: No
 
 ## Utils
