+++ projects/plugins/jetpack/extensions/shared/components/media-player-control/index.js
@@ -7,30 +7,48 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { ToolbarGroup, ToolbarButton } from '@wordpress/components';
+import { useSelect, useDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
 import './style.scss';
 import { ControlBackFiveIcon, ControlForwardFiveIcon } from '../../icons';
-import { STATE_PAUSED, STATE_PLAYING } from '../../../store/media-source/constants';
+import { STATE_PAUSED, STATE_PLAYING, STORE_ID } from '../../../store/media-source/constants';
 import { convertSecondsToTimeCode } from './utils';
 
 function noop () {};
 
 export default function MediaPlayerControl( {
-	time,
-	state = STATE_PAUSED,
 	skipForwardTime = 5,
 	jumpBackTime = 5,
 	playIcon = 'controls-play',
 	pauseIcon = 'controls-pause',
-	isDisabled = false,
-
 	onTimeChange = noop,
-	onToggle = noop,
 } ) {
-	const timeInFormat = convertSecondsToTimeCode( time );
+	const {
+		playerState,
+		playerCurrentTime,
+		defaultMediaSource,
+	} = useSelect( select => {
+		const {
+			getMediaSourceCurrentTime,
+			getMediaPlayerState,
+			getDefaultMediaSource,
+		} = select( STORE_ID );
+
+		return {
+			playerState: getMediaPlayerState(),
+			playerCurrentTime: getMediaSourceCurrentTime(),
+			defaultMediaSource: getDefaultMediaSource(),
+		}
+	}, [] );
+
+	const timeInFormat = convertSecondsToTimeCode( playerCurrentTime );
+	const isDisabled = ! defaultMediaSource;
+
+	const { toggleMediaSource } = useDispatch( STORE_ID );
+	const togglePlayer = () => toggleMediaSource( defaultMediaSource.id );
 
 	return (
 		<ToolbarGroup>
@@ -38,27 +56,21 @@ export default function MediaPlayerControl( {
 				<ToolbarButton
 					icon={ ControlBackFiveIcon }
 					isDisabled={ isDisabled }
-					onClick={ () => onTimeChange( time - jumpBackTime ) }
+					onClick={ () => onTimeChange( playerCurrentTime - jumpBackTime ) }
 				/>
 			) }
 			
 			<ToolbarButton
-				icon={ state === STATE_PAUSED ? playIcon : pauseIcon }
+				icon={ playerState === STATE_PAUSED ? playIcon : pauseIcon }
 				isDisabled={ isDisabled }
-				onClick={ () => {
-					if ( state === STATE_PLAYING ) {
-						return onToggle && onToggle( STATE_PAUSED );
-					}
-					
-					return onToggle( STATE_PLAYING );
-				} }
+				onClick={ togglePlayer }
 			/>
 
 			{ skipForwardTime && (
 				<ToolbarButton
 					icon={ ControlForwardFiveIcon }
 					isDisabled={ isDisabled }
-					onClick={ () => onTimeChange( time + skipForwardTime ) }
+					onClick={ () => onTimeChange( playerCurrentTime + skipForwardTime ) }
 				/>
 			) }
